fixes for threads and data types

+++ src/main.cpp
@@ -2,6 +2,7 @@
 #include <chrono>
 #include <iostream>
 #include <mutex>
+#include <pthread.h>
 #include <queue>
 #include <thread>
 
@@ -87,7 +88,7 @@ int main(int argc, char** argv) {
 	}
 	// wait for threads
 	reader.join();
-	for (int i = 0; i < parsers.size(); ++i) {
+	for (size_t i = 0; i < parsers.size(); ++i) {
 		parsers[i].join();
 	}
 	results.print_errors();
